updated onboard js, refactored
